Uniform use of splitHTMLColor
