+++ src/main/java/com/glencoesoftware/omero/ms/image/region/ImageRegionCtx.java
@@ -367,50 +367,4 @@ public class ImageRegionCtx extends OmeroRequestCtx {
             // Ignore
         }
     }
-
-    /**
-     *  Splits an hex stream of characters into an array of bytes
-     *  in format (R,G,B,A) and converts to a
-     *  {@link ome.xml.model.primitives.Color}.
-     *  - abc      -> (0xAA, 0xBB, 0xCC, 0xFF)
-     *  - abcd     -> (0xAA, 0xBB, 0xCC, 0xDD)
-     *  - abbccd   -> (0xAB, 0xBC, 0xCD, 0xFF)
-     *  - abbccdde -> (0xAB, 0xBC, 0xCD, 0xDE)
-     *  @param color Characters to split.
-     *  @return corresponding {@link ome.xml.model.primitives.Color}
-     */
-    public static Integer splitHTMLColor(String color) {
-        List<Integer> level1 = Arrays.asList(3, 4);
-        try {
-            if (level1.contains(color.length())) {
-                String c = color;
-                color = "";
-                for (char ch : c.toCharArray()) {
-                    color += ch + ch;
-                }
-            }
-            if (color.length() == 6) {
-                color += "FF";
-            }
-            if (color.length() == 8) {
-                int r = Integer.parseInt(color.substring(0, 2), 16);
-                int g = Integer.parseInt(color.substring(2, 4), 16);
-                int b = Integer.parseInt(color.substring(4, 6), 16);
-                int a = Integer.parseInt(color.substring(6, 8), 16);
-                Color clr = new Color(r, g, b, a);
-                return clr.getValue();
-            }
-        } catch (Exception e) {
-            log.error("Error while parsing color: {}", color, e);
-        }
-        return null;
-    }
-
-    public List<Color> omeColors() {
-        List<Color> clrs = new ArrayList<Color>();
-        for(String c : colors) {
-            clrs.add(new Color(splitHTMLColor(c)));
-        }
-        return clrs;
-    }
 }
