+++ src/main/java/com/glencoesoftware/omero/ms/image/region/OmeroRenderingHandler.java
@@ -1,21 +1,3 @@
-/*
- * Copyright (C) 2021 Glencoe Software, Inc. All rights reserved.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
- */
-
 package com.glencoesoftware.omero.ms.image.region;
 
 import java.util.HashMap;
@@ -48,10 +30,10 @@ import omero.api.IQueryPrx;
 import omero.sys.ParametersI;
 import omero.util.IceMapper;
 
-public class OmeroRequestHandler {
+public class OmeroRenderingHandler extends OmeroVertxRequestHandler{
 
     private static final org.slf4j.Logger log =
-            LoggerFactory.getLogger(OmeroRequestHandler.class);
+            LoggerFactory.getLogger(OmeroRenderingHandler.class);
 
     /** Reference to the compression service. */
     protected final LocalCompress compressionSrv;
@@ -74,7 +56,7 @@ public class OmeroRequestHandler {
     /** Configured OmeroZarrUtils */
     protected final OmeroZarrUtils zarrUtils;
 
-    public OmeroRequestHandler(
+    public OmeroRenderingHandler(
             LocalCompress compressionSrv,
             LutProvider lutProvider,
             List<Family> families,
